pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "fraudguard-ml"
version = "2.1.0"
description = "Production-grade real-time transaction fraud detection system"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "ML Platform Team", email = "ml-platform@company.com" }]

dependencies = [
    # Web framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # ML core
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.3",
    "torch>=2.1.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",

    # Experiment tracking & model registry
    "mlflow>=2.10.0",
    "optuna>=3.5.0",

    # Data validation
    "great-expectations>=0.18.0",

    # Feature store / caching
    "redis>=5.0.1",
    "hiredis>=2.3.2",

    # Streaming
    "aiokafka>=0.10.0",
    "confluent-kafka>=2.3.0",

    # Monitoring
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",

    # Data versioning
    "dvc[s3]>=3.38.0",

    # Utilities
    "structlog>=24.1.0",
    "rich>=13.7.0",
    "httpx>=0.26.0",
    "tenacity>=8.2.3",
    "python-multipart>=0.0.7",
    "python-jose[cryptography]>=3.3.0",
    "imbalanced-learn>=0.11.0",
    "shap>=0.44.0",
    "scipy>=1.12.0",
    "joblib>=1.3.2",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "factory-boy>=3.3.0",
    "httpx>=0.26.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "ipykernel>=6.29.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]

[project.scripts]
fraudguard-train = "scripts.run_training:main"
fraudguard-serve = "src.api.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "ANN", "S", "B", "A", "COM", "C4", "DTZ", "EM", "ISC", "PIE", "T20", "SIM", "TCH", "ARG", "PTH"]
ignore = ["ANN101", "ANN102", "S101", "COM812", "ISC001"]

[tool.ruff.per-file-ignores]
"tests/*" = ["ANN", "S", "ARG"]
"scripts/*" = ["T20"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-report=xml -v"
markers = [
    "unit: Unit tests (fast, no external services)",
    "integration: Integration tests (requires Docker services)",
    "slow: Slow tests (>30s)",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
fail_under = 80
